+++ projects/angular-adapter/src/lib/interfaces/index.ts
@@ -1,9 +1,2 @@
-import { FlatfileSettings } from './settings';
-
-export * from './angular-adapter.config';
-export {
-  RecordInitOrChangeCallback,
-  ScalarDictionaryWithCustom,
-} from './general';
-export * from './row-record.interface';
-export { FlatfileSettings };
+export * from './output-method.interface';
+export * from './params.interface';