+++ projects/angular-adapter/src/lib/interfaces/row-record.interface.ts
@@ -1,6 +1,0 @@
-import { ScalarDictionaryWithCustom } from './general';
-
-export interface RowRecord {
-  data: ScalarDictionaryWithCustom;
-  index: number;
-}