+++ projects/angular-adapter/src/public-api.ts
@@ -2,21 +2,14 @@
  * @note Public API Surface of angular-adapter
  */
 
-import FlatfileImporter, {
-  FieldHookCallback,
-  CustomerObject as FlatfileCustomer,
-  FlatfileResults,
-  IDataHookResponse,
-} from '@flatfile/adapter';
+import {
+  flatfileImporter,
+  IFlatfileImporter,
+  IUnsafeGenerateTokenOptions,
+} from '@flatfile/sdk';
 
 export * from './lib/flatfile-adapter.module';
 export * from './lib/flatfile-button.component';
 export * from './lib/interfaces';
 
-export {
-  FlatfileCustomer,
-  FlatfileImporter,
-  FlatfileResults,
-  FieldHookCallback,
-  IDataHookResponse,
-};
+export { flatfileImporter, IFlatfileImporter, IUnsafeGenerateTokenOptions };